Single executable
Collapse scripts into a single script with sub-commands
