+++ rpm/deployment-support.spec.in
@@ -28,9 +28,7 @@ rm -f requirements.txt
 %doc README.rst
 %{python2_sitelib}/legacy_migration
 %{python2_sitelib}/deployment_support-%%{version}*.egg-info
-%{_bindir}/cfg_mgmt_tool
-%{_bindir}/toggle_cfg
-%{_bindir}/update_domain
+%{_bindir}/deployment_tool
 
 %changelog
 * Mon Nov 26 2018 Thomas Bachman <tbachman@yahoo.com> - 1.0.0-1
